idlestat: make energy model template file
If the energy model file doesn't exist, instead
make an energy model template file (taking the
topology and C and P-state names/values into account).
This file is similar to an energy model file, except that
all of the energy costs are to be filled in (and are shown
as "?").

Signed-off-by: Larry Bassel <larry.bassel@linaro.org>
Reviewed-by: Tuukka Tikkanen <tuukka.tikkanen@linaro.org>