Energy_model: Do not fail when no energy model is used
Previous patchset unconditionally tries to load energy model, even
if the energy model file is not specified (i.e. energy model
processing is not desired). This patch modifies the code to skip
parsing energy model in case it is not specified.

The call to parsing energy model is moved from getoptions() to main().

Signed-off-by: Tuukka Tikkanen <tuukka.tikkanen@linaro.org>
Reviewed-by: Larry Bassel <larry.bassel@linaro.org>