idlestat: Replace asserts with proper error handling
Some parts of the code use asserts as a way to terminate execution
in case input data processing fails. This leaves the user in the black
regarding the reason why the tool terminated (and possibly dumped a
core file).

Replace the abnormal termination with displaying error message and
terminating the application in a controlled fashion.

Replace asserts with displaying error messages and/or terminating
program in a controlled manner. Also refine some error handling paths.

Leave asserts that are triggered by program internal behavior intact.

Signed-off-by: Pi-Cheng Chen <pi-cheng.chen@linaro.org>
Signed-off-by: Tuukka Tikkanen <tuukka.tikkanen@linaro.org>