idlestat: Suppress invalid warnings for generated P-state changes
An earlier patch introduced generating P-state changes into
beginning and end of the trace file. This was done to account
P-state residencies that extended outside of the trace period
correctly. These P-state changes reliably trigger warning
messages about changing a P-state for an idle CPU. (These could
also be triggered by P-state changes that occur naturally.)

This patch suppresses warnings about P-state changes for "an idle
CPU" when the actual idle state of the CPU is not known.

Signed-off-by: Tuukka Tikkanen <tuukka.tikkanen@linaro.org>

Reviewed-by: Amit Kucheria <amit.kucheria@linaro.org>