idlestat: Display error messages closer to point of failure
Functions build_cstate_info and load_and_build_cstate_info
previously handled errors by returning an error code without
printing an error message. This was a problem because the error
could be caused by something which does not set errno, but
the caller assumed this was always the case and called perror()
to report the error to user.

This patch adds reporting of errors directly from these
functions and changes the caller error propagation to be silent.

Signed-off-by: Tuukka Tikkanen <tuukka.tikkanen@linaro.org>

Reviewed-by: Amit Kucheria <amit.kucheria@linaro.org>