idlestat: Fix recording of initial and final P-states
The initial and final pstates were stored only if the display
options at the time of creating the trace file included display
of P-states. Since the trace may later be imported with the
P-state display on, add the pseudo P-state changes to the trace
unconditionally.

In addition, the return value of get_trace_ts was never checked
for error. The error checking has been added. Since floating
point comparison for equality is unreliable, the function now
returns only status and the actual data is relayed by modifying
memory pointed to by a function parameter. Error messages have
been added to the get_trace_ts function.

Signed-off-by: Tuukka Tikkanen <tuukka.tikkanen@linaro.org>

Reviewed-by: Amit Kucheria <amit.kucheria@linaro.org>