idlestat: Make ptrerror return a non-NULL pointer
Some functions such as inter() and intersection() use NULL return
value to indicate both errors (e.g. out of memory) and a result
(empty intersection). Make ptrerror() return a unique non-NULL
pointer value so that the error and NULL result can be told apart.

Previously the parameter to ptrerror() was unconditionally
interpreted as a pointer to string to be passed to perror(). Now
the function checks the value and calls perror only if the
parameter value is not NULL.

A new function, is_err(), has been added to check if any pointer
value is equal to ptrerror() return value.

error(), ptrerror() and is_err() have been moved to utils.c.

The functions that used ptrerror() before this patch have been
modified to consistently return ptrerror() instead of NULL in
cases of errors (affects the return statements where no call to
perror() is desired). Likewise the callers now use is_err()
instead of checking for NULL pointer.

Signed-off-by: Tuukka Tikkanen <tuukka.tikkanen@linaro.org>