idlestat: Make allocations consistent and safer
This patch modifies malloc/calloc/realloc calls in idlestat.c to
conform to following rules:
1. Newly allocated memory is zeroed by using calloc instead of malloc
or by calling memset after realloc. The only exception is when the newly
allocated memory is storage for a string or always completely overwritten
with a call to memcpy or a single assignment statement.
2. Calls to memset(ptr, 0, size) that followed calloc have been removed.
3. perror() message body after allocation failure is uniformly __func__,
resulting in "func_name: Out of memory".

Signed-off-by: Tuukka Tikkanen <tuukka.tikkanen@linaro.org>