topology: Dynamic allocation for topologies
Pre-patch the topology is rooted to a global variable g_cpu_topo_list.
This prevents loading multiple traces for comparison. This patch
makes the allocation dynamic. The root pointer is stored within
struct cpuidle_datas, allowing the idlestat_load() to retain its
prototype, but most of the functions defined in topology.c have
a new parameter for passing the topology root pointer.

Signed-off-by: Tuukka Tikkanen <tuukka.tikkanen@linaro.org>

Reviewed-by: Koan-Sin Tan <freedom.tan@linaro.org>