idlestat: Rename print_vrb and make it a real function
print_vrb (print verbose) is a preprocessor macro defined in utils.h.
It behaves like fprintf(stderr, ...) if the level argument is greater
than or equal to current options->verbose value (incremented by one
for each -v parameter on command line). Unfortunately the macro does
depend on variable "options" being visible in the scope it is used in,
which prevents using it in any function that does not have that
variable declared in it or passed to it.

This patch makes the macro a real function and disconnects the
dependency of having options visible at call site. Instead while the
options are passed, a call to set_verbose_level() is make whenever
the value of options->verbose changes.

The name and call semantics are adjusted to increase versatility of
the function. Two variants are created:
verbose_printf(int min_level, const char *fmt, ...)
verbose_fprintf(FILE *f, int min_level, const char *fmt, ...)
These two behave like printf and fprintf if level >= current verbose
level, otherwise they do nothing and return 0.

Signed-off-by: Tuukka Tikkanen <tuukka.tikkanen@linaro.org>