report_ops: Add mechanism for report_ops to check program options
Some reports may require some program options to be set, but other
reports may not require the same options. To keep details of these
checks out of main() and getoptions(), this patch adds a new
operation to report_ops for doing these checks when necessary.

Pointer to struct program_options is passed to the op. If the
options are not suitable for operation, the op must print an error
message for the user and return -1. Otherwise the op should
return 0.

Signed-off-by: Tuukka Tikkanen <tuukka.tikkanen@linaro.org>