idlestat: Add support for baseline datas
This patch adds command line option for specifying a baseline
trace file. If specified, the trace is loaded using idlestat_load
and the baseline member in the main datas structure points to
the loaded baseline datas structure. As such, this adds no new
features itself, but allows implementing a comparison feature.

Signed-off-by: Tuukka Tikkanen <tuukka.tikkanen@linaro.org>