Reports: Add support for allocating and releasing report_data
While passing around the per report data pointer has been previously
implemented, there was no mechanism for the report ops to allocate
or release the data as needed. This patch moves allocation of
report_data pointer from struct options to main() and adds calls
necessary for allocating (and initializing) and releasing the data.

Signed-off-by: Tuukka Tikkanen <tuukka.tikkanen@linaro.org>