Idlestat: Prevent alloc_pstate from creating duplicate entries
The existing version of alloc_pstate() can create duplicate
entries for the same pstate if called with a frequency
value that already exists. Change the semantics so that
in case the frequency already exists in the array, the
index returned is for the existing frequency.

This change allows removal of (now) obsolete function
get_current_pstate and merging of pstate entries between
main trace and baseline trace for comparison report purposes.

Signed-off-by: Tuukka Tikkanen <tuukka.tikkanen@linaro.org>