Topology: Add functions for setting up loadtime cluster/core data
Current implementation for core/cluster data calculation is to look
for intersections of same state residency in cpu data after loading
all of the data. This fails to correctly record core/cluster data
for cases where all the contained cpus are in a low power state, but
are not using the exact same state.

To resolve this, core/cluster state needs to be evaluated at the
same time as the cpu events are being processes. This patch adds
functions for allocating and initializing memory for core/cluster
statistics.

Signed-off-by: Tuukka Tikkanen <tuukka.tikkanen@linaro.org>