Idlestat: Fix core/cluster cstates
Calls to core_get_least_cstate and cluster_get_least_cstate
were not included in the patch that intended to add them. This
caused all core/cluster c-states to be recorded as if the
core/cluster entered exactly the same state as what was
entered by any cpu in the core/cluster that changed its state.

Signed-off-by: Tuukka Tikkanen <tuukka.tikkanen@linaro.org>