Idlestat: Add command line options for managing trace buffer
As the kernel trace buffer is a finite resource, the default
trace buffer size idlestat sets may either be impossibly
large or insufficient to hold the data. This patch adds two
new command line options to control buffer size and whether
or not it will be polled during the trace period.

The polling functionality while tracing is not yet implemented.

Signed-off-by: Tuukka Tikkanen <tuukka.tikkanen@linaro.org>