Energy_model: Don't add phony frequency to template
Idlestat adds phony frequency entries to the end of trace file
to ensure all the last running state is accounted for. This
phony frequency was unsigned int "-1" (aka 2^32-1 or 2^64-1),
which was then listed in any energy model template generated
from the trace file.

This patch changes the phony frequency to 0 and makes the
energy model template generation ignore 0 frequency.

Signed-off-by: Tuukka Tikkanen <tuukka.tikkanen@linaro.org>