Remove unused parameter cpu from write_cstate_info
The parameter cpu is not used in write_cstate_info.

Note: The function is intended as an abstraction for writing
cstate information to trace file. As the trace file format
may change in future or multiple output formats may be
forthcoming, this function is an abstraction for storing
the state. Some future format may require the cpu number,
so this parameter may be reintroduced later. However, as
we currently do not know whether it is needed or not, it
can be removed.

Signed-off-by: Tuukka Tikkanen <tuukka.tikkanen@linaro.org>