Add -Wunused-parameter to CFLAGS
Android complation checks that all parameters are used. Add
check for this condition to default compilation as well. For
cases where the parameter is legitimately not used, add a
macro UNUSED to flag such cases for the compiler. This macro
is defined in a new header file. Currently this macro is
defined functional if the preprocessor macro __GNUC__ is
defined, otherwise the macro expands to an empty comment.

Signed-off-by: Tuukka Tikkanen <tuukka.tikkanen@linaro.org>